+++ smol/cofe/__init__.py
@@ -7,10 +7,6 @@ functions to define and fit cluster expansions for crystalline materials.
 
 from .space.clusterspace import ClusterSubspace, PottsSubspace
 from .expansion import ClusterExpansion
-from smol.cofe.wrangling.wrangler import (StructureWrangler,
-                                          weights_energy_above_composition,
-                                          weights_energy_above_hull)
+from smol.cofe.wrangling.wrangler import StructureWrangler
 
-__all__ = ['ClusterSubspace', 'PottsSubspace', 'StructureWrangler',
-           'ClusterExpansion', 'weights_energy_above_composition',
-           'weights_energy_above_hull']
+__all__ = ['ClusterSubspace', 'StructureWrangler', 'ClusterExpansion']