+++ smol/cofe/wrangling/wrangler.py
@@ -16,55 +16,13 @@ composition for a given set of structures.
 __author__ = "Luis Barroso-Luque"
 __credits__ = "William Davidson Richard"
 
+from typing import Sequence
 import warnings
-from collections import defaultdict
 import numpy as np
 from monty.json import MSONable
-from pymatgen import Composition, Structure
-from pymatgen.analysis.phase_diagram import PhaseDiagram, PDEntry
+from pymatgen import Structure
 from smol.cofe.space.clusterspace import ClusterSubspace
 from smol.exceptions import StructureMatchError
-from smol.constants import kB
-
-
-# TODO move these to a separate toolbox file
-def weights_energy_above_composition(structures, energies, temperature=2000):
-    """Compute weights for energy above the minimum reduced composition energy.
-
-    Args:
-        structures (list):
-            list of pymatgen.Structures
-        energies (ndarray):
-            energies of corresponding structures.
-        temperature (float):
-            temperature to used in boltzmann weight
-
-    Returns: weights for each structure.
-        array
-    """
-    e_above_comp = _energies_above_composition(structures, energies)
-    return np.exp(-e_above_comp / (kB * temperature))
-
-
-def weights_energy_above_hull(structures, energies, cs_structure,
-                              temperature=2000):
-    """Compute weights for structure energy above the hull of given structures.
-
-    Args:
-        structures (list):
-            list of pymatgen.Structures
-        energies (ndarray):
-            energies of corresponding structures.
-        cs_structure (Structure):
-            The pymatgen.Structure used to define the cluster subspace
-        temperature (float):
-            temperature to used in boltzmann weight.
-
-    Returns: weights for each structure.
-        array
-    """
-    e_above_hull = _energies_above_hull(structures, energies, cs_structure)
-    return np.exp(-e_above_hull / (kB * temperature))
 
 
 class StructureWrangler(MSONable):
@@ -101,6 +59,7 @@ class StructureWrangler(MSONable):
         """
         self._subspace = cluster_subspace
         self._items = []
+        self._ind_sets = {}  # data indices for test/training splits etc
         self._metadata = {'applied_filters': []}
 
     @property
@@ -123,6 +82,11 @@ class StructureWrangler(MSONable):
         """Get list of properties that have been added."""
         return list(set(p for i in self._items
                         for p in i['properties'].keys()))
+
+    @property
+    def available_indices(self):
+        """Get list of available data index sets."""
+        return list(self._ind_sets.keys())
 
     @property
     def available_weights(self):
@@ -304,6 +268,22 @@ class StructureWrangler(MSONable):
             properties /= self.sizes
 
         return properties
+
+    def data_indices(self, key):
+        """Get a specific data index set."""
+        return self._ind_sets[key]
+
+    def add_data_indices(self, key, indices):
+        """Add a set of data indices.
+
+        Fore example use this for saving test/training splits or separating
+        duplicates.
+        """
+        if not isinstance(indices, (Sequence, np.ndarray)):
+            raise TypeError("indices must be Sequence like or an ndarray.")
+        elif any(i not in range(self.num_structures) for i in indices):
+            raise ValueError("One or more indices are out of range.")
+        self._ind_sets[key] = list(indices)
 
     def get_weights(self, key):
         """Get the weights specified by the given key.
@@ -601,6 +581,7 @@ class StructureWrangler(MSONable):
                           'weights': item['weights']})
         sw._items = items
         sw._metadata = d['metadata']
+        sw._ind_sets = d.get('_ind_sets') or {}
         return sw
 
     def as_dict(self):
@@ -624,48 +605,8 @@ class StructureWrangler(MSONable):
              '@class': self.__class__.__name__,
              '_subspace': self._subspace.as_dict(),
              '_items': s_items,
+             '_ind_sets': self._ind_sets,
              'metadata': self.metadata}
         return d
 
 
-def _energies_above_hull(structures, energies, ce_structure):
-    """Compute energies above hull.
-
-    Hull is constructed from phase diagram of the given structures.
-    """
-    pd = _pd(structures, energies, ce_structure)
-    e_above_hull = []
-    for s, e in zip(structures, energies):
-        entry = PDEntry(s.composition.element_composition, e)
-        e_above_hull.append(pd.get_e_above_hull(entry))
-    return np.array(e_above_hull)
-
-
-def _pd(structures, energies, cs_structure):
-    """Generate a phase diagram with the structures and energies."""
-    entries = []
-
-    for s, e in zip(structures, energies):
-        entries.append(PDEntry(s.composition.element_composition, e))
-
-    max_e = max(entries, key=lambda e: e.energy_per_atom).energy_per_atom
-    max_e += 1000
-    for el in cs_structure.composition.keys():
-        entry = PDEntry(Composition({el: 1}).element_composition, max_e)
-        entries.append(entry)
-
-    return PhaseDiagram(entries)
-
-
-def _energies_above_composition(structures, energies):
-    """Compute structure energies above reduced composition."""
-    min_e = defaultdict(lambda: np.inf)
-    for s, e in zip(structures, energies):
-        comp = s.composition.reduced_composition
-        if e / len(s) < min_e[comp]:
-            min_e[comp] = e / len(s)
-    e_above = []
-    for s, e in zip(structures, energies):
-        comp = s.composition.reduced_composition
-        e_above.append(e / len(s) - min_e[comp])
-    return np.array(e_above)