+++ CHANGES.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Use this section to keep track of changes in the works.
 
 ### Added
+* data indices in `StructureWrangler` to keep track of training/test splits,
+  duplicate sets, etc.
+* `ClusterSubspace.cutoffs` property to obtain tight cutoffs of included
+   orbits.
 * Added properties to get orbit and ordering multiplicities of corr functions.
 [\#102](https://github.com/CederGroupHub/smol/pull/102)
 ([lbluque](https://github.com/lbluque))
@@ -20,6 +24,9 @@ Use this section to keep track of changes in the works.
 
 
 ### Changed
+* filter functions in `smol.wrangling` replaced with functions returning
+  indices corresponding to structures to keep. This can be used saving indices
+  with `StructureWrangler.add_data_indices`.
 * Cleanup of sites, active sites and restricted sites in `Sublattice`
 [\#95](https://github.com/CederGroupHub/smol/pull/95)
   ([juliayang](https://github.com/juliayang))